% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-meta-package.R
\name{create_meta_package}
\alias{create_meta_package}
\title{Create a 'Verse' or 'Meta' Package}
\usage{
create_meta_package(name, imports, dev)
}
\arguments{
\item{name}{The meta package name}

\item{imports}{A character vector of the packages you're bundling}

\item{dev}{If any of the packages included are development versions (e.g. on
GitHub and not on CRAN), supply a vector of TRUE/FALSE indicating the
development version should be used.}
}
\description{
Inspired by the package \href{https://tidyverse.tidyverse.org/}{tidyverse}, this
function helps you easily create your own 'verse' or 'Meta' package to
facilitate easy installation and loading of multiple, related packages. Plus,
you'll get a nice 'tidyverse' style startup message.
}
\examples{

\dontrun{
create_meta_package(name = "myPackageName",
                    imports = c("publishedPackage1",
                                "publishedPackage2")
)


create_meta_package(name = "myPackageName",
                    imports = c("publishedPackage1",
                                "devPackage",
                                "publishedPackage2",
                    ),
                    dev = c(FALSE,
                            TRUE,
                            FALSE)
)
}

}
